[REFACTOR] : ♻️ react-hook-form 사용 후 미사용 코드 대폭 제거 및, rename
